[CHORE]: Refactorización en backend para las peticiones de contribuciones #37
